LLM/Project：RAG PDF/requirements.txt
streamlit
chromadb
openai
python-dotenv
PyPDF2
uuid
